Refactor LocalUser Components to be in Component folder
