+++ src/components/Localuser/Localuser.tsx
@@ -1,13 +1,13 @@
 import React, { useRef } from "react"
 import styled from "styled-components"
 import { useLocalStore } from "../../store/LocalStore"
-import LocalVideo from "./LocalVideo"
-import LocalAudio from "./LocalAudio"
-import { MuteIndicator } from "./elements/MuteIndicator"
+import LocalVideo from "./components/LocalVideo"
+import LocalAudio from "./components/LocalAudio"
+import { MuteIndicator } from "./components/MuteIndicator"
 import { ReloadHint } from "../ReloadHint/ReloadHint"
 import { panOptions } from "./../PanWrapper/panOptions"
-import { AudioRadius } from "./elements/AudioRadius"
-import { NameContainer } from "./elements/NameContainer"
+import { AudioRadius } from "./components/AudioRadius"
+import { NameContainer } from "./components/NameContainer"
 
 const Container = styled.div`
   width: ${panOptions.user.size.x}px;
